refactor: Simplify WorkLogCalendar and enhance WorkLogSection layout
- Removed unused month formatting and month change functionality from WorkLogCalendar.
- Updated WorkLogSection to include WorkLogMonthPicker for improved month selection.
- Adjusted layout in WorkLogSection for better alignment and responsiveness.
